Create package github with oauth logic
